+++ rentmanager/rentmanager/src/main/java/com/epf/rentmanager/service/ClientService.java
@@ -40,7 +40,7 @@ public class ClientService {
 		}
 	}
 
-	public Client findById(long id) throws ServiceException {
+	public Client findById(int id) throws ServiceException {
 		try {
 			return clientDao.findById(id);
 		} catch (DaoException e) {
